// digital-money-house/src/types/home.types.ts
export interface HomeData {
    _id: string;
    altDesktop: string;
    altLogo: string;
    altMobile: string;
    descripcionServicio1: string;
    descripcionServicio2: string;
    servicio1: string;
    servicio2: string;
    subtitulo: string;
    titulo: string;
    urlDesktop: string;
    urlLogo: string;
    urlMobile: string;
}